pyproject.toml
[project]
name = "langgraph-reflection"
version = "0.0.1"
description = "LangGraph agent that runs a reflection step"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph",
    "mypy>=1.8.0",
    "langchain>=0.1.0",
    "python-dotenv>=1.0.0"
]
authors = [{name = "Harrison Chase"}]

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "langgraph-reflection",
]
